button directive moved inside directives folder
