+++ src/main.ts
@@ -20,7 +20,7 @@ export * from "./grid/";
 
 export * from "./navigation-drawer/navigation-drawer.component";
 export * from "./icon/icon.component";
-export * from "./button/button.directive";
+export * from "./directives/button/button.directive";
 export * from "./buttonGroup/buttonGroup.component";
 export * from "./list/list.component";
 export * from "./tabbar/tabbar.component";
